feat: add helper functions
